Add support for prefixed names in bibkey formatter
Prefixed names like for instance 'Van Hove' were not treated properly, i.e.
it was understood as one name such that the resulting key contained a white
space. The implemented mechanism now allows to treat prefixed names properly
by splitting author names at white spaces and apply the formatter to the
name and prefix separately. As an example consider a key format which contains
the capitalized first author name: Previously for prefixed names this would
have led to a key of the form 'Van hove' which is now transformed the format
'VanHove'