Use collections.OrderedDict to store file contents
Prior to this change a default dictionary was used to store the field
names together with the field contents for bib-file entries. For python
version < 3.6 (dictionary input order not preserved) this lead to random
order of bibtex entries in the output file. Although, not of any practical
importance for the bibtex files this did not allow to run test based on
simple file comparisons.