make content splitting algorithm more robust
In the previous implementation single BibTeX entries were assumed to be
separated by ,\n and were split accordingly. Since for instance abstract
field are likely to contain similar strings this methods may lead to erroneous
splitting of content entries.

To avoid this issue in the future the algorithm used to identify single
entries was updated and now splits the contents and additionally checks that
curly braces in entries resulting from the splitting are balanced.